reorder files, new pages, img and styles folders, removed unused/redundant components, renamed components to english
