+++ src/App.js
@@ -1,14 +1,14 @@
 import React from "react";
-import "./App.css";
+import "./styles/App.css";
 import { Route, Routes } from "react-router-dom";
-import LogIn from "./Components/LogIn";
-import SignUp from "./Components/SignUp.js";
-import Reserva from "./Components/Reserva";
+import LogIn from "./Pages/LogInPage.js";
+import SignUp from "./Pages/RegisterPage.js";
+import Reserva from "./Pages/ReservationsPage.js";
 import ProfileMenu from "./Components/ProfileMenu";
 import LoggedOutHeader from "./Components/LoggedOutHeader.js";
-import MovieList from "./Utils/MovieList.js";
-import MovieUpload from "./Components/MovieUpload.js";
-import Home from "./Components/Home.js";
+import MovieList from "./Components/MovieList.js";
+import MovieUpload from "./Pages/MovieUploadPage.js";
+import Home from "./Pages/HomePage.js";
 import { Link } from "react-router-dom";
 import logo from "./logo.jpeg";
 import Button from "./Components/Button.js";
