--- src/Components/FrontPage.css
@@ -1,16 +1,0 @@
-.main-banner {
-    background-size: cover;
-    background-position: center;
-    height: 500px;
-    position: relative;
-}
-
-.movie-info {
-    position: absolute;
-    bottom: 20px;
-    left: 20px;
-    background-color: rgba(0, 0, 0, 0.5);
-    color: white;
-    padding: 10px;
-    border-radius: 5px;
-}
