+++ src/Components/LoggedOutHeader.js
@@ -1,6 +1,5 @@
 import { Link } from "react-router-dom";
 import Button from "./Button";
-import "./Button.css";
 
 const LoggedOutHeader = () => {
   return (
