+++ src/Pages/HomePage.js
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import Button from "./Button";
+import Button from "../Components/Button";
 
 const Home = () => {
   return (
