+++ src/styles/Listings.css
